feat: params from least square seems fine

+++ src/linear_regression/utils.c
@@ -5,9 +5,9 @@ float square(float n)
     return n * n;
 }
 
-float mean(float *n, int len)
+float mean(int *n, int len)
 {
-    float sum = 0;
+    int sum = 0;
     for(int i = 0; i < len; i ++){
         sum += n[i];
     }
